namada_storage::collections
Module lazy_map

Lazy map.

Structs

LazyMap

Lazy map.

Enums

NestedSubKey

Possible sub-keys of a nested LazyMap

SubKey

Possible sub-keys of a LazyMap

ValidationError

Constants

DATA_SUBKEY

Subkey corresponding to the data elements of the LazyMap

Traits

Collectable

Trait used to facilitate collection of lazy maps into eager maps

Type Aliases

NestedMap

A LazyMap with another LazyCollection inside it’s value V