namada_storage::collections::lazy_map
Trait Collectable

pub trait Collectable {
    type Collected;

    // Required method
    fn collect_map<S: StorageRead>(
        &self,
        storage: &S,
    ) -> Result<Self::Collected>;
}

Trait used to facilitate collection of lazy maps into eager maps

Required Associated Types

type Collected

The type of the value of the lazy map

Required Methods

fn collect_map<S: StorageRead>(&self, storage: &S) -> Result<Self::Collected>

Collect the lazy map into an eager map

Object Safety

This trait is not object safe.